feat: split types into their own package
As part of the scope analysis work, I want to remove its runtime dependency on `typescript-estree`, so it doesn't need to depend on `typescript`.
It only needs a few pieces from the package; AST_NODE_TYPES, and the visitor keys.

This moves the following the to a new package, `@typescript-eslint/types`:
- `AST_NODE_TYPES`
- `AST_TOKEN_TYPES`
- `TSESTree` (aka the types for the typescript-estree AST)
- `ParserOptions` (moved out of `experimental-utils`)

This has the added bonus of getting us very close to removing the need for `parser` to depend on `experimental-utils`, which will reduce its dependency tree significantly.